+++ beetsplug/fetchart.py
@@ -1,5 +1,5 @@
 # This file is part of beets.
-# Copyright 2011, Adrian Sampson.
+# Copyright 2012, Adrian Sampson.
 #
 # Permission is hereby granted, free of charge, to any person obtaining
 # a copy of this software and associated documentation files (the
@@ -12,21 +12,24 @@
 # The above copyright notice and this permission notice shall be
 # included in all copies or substantial portions of the Software.
 
-"""Finding album art for tagged albums.
+"""Fetches album art.
 """
 import urllib
+import re
 import logging
 import os
-import re
+
+from beets.plugins import BeetsPlugin
 
 IMAGE_EXTENSIONS = ['png', 'jpg', 'jpeg']
 COVER_NAMES = ['cover', 'front', 'art', 'album', 'folder']
+CONTENT_TYPES = ('image/jpeg',)
 
-# The common logger.
 log = logging.getLogger('beets')
 
 
-CONTENT_TYPES = ('image/jpeg',)
+# ART SOURCES ################################################################
+
 def _fetch_image(url):
     """Downloads an image from a URL and checks whether it seems to
     actually be an image. If so, returns a path to the downloaded image.
@@ -50,7 +53,8 @@ def _fetch_image(url):
 # Art from Amazon.
 
 AMAZON_URL = 'http://images.amazon.com/images/P/%s.%02i.LZZZZZZZ.jpg'
-AMAZON_INDICES = (1,2)
+AMAZON_INDICES = (1, 2)
+
 def art_for_asin(asin):
     """Fetches art for an Amazon ID (ASIN) string."""
     for index in AMAZON_INDICES:
@@ -65,6 +69,7 @@ def art_for_asin(asin):
 
 AAO_URL = 'http://www.albumart.org/index_detail.php'
 AAO_PAT = r'href\s*=\s*"([^>"]*)"[^>]*title\s*=\s*"View larger image"'
+
 def aao_art(asin):
     # Get the page from albumart.org.
     url = '%s?%s' % (AAO_URL, urllib.urlencode({'asin': asin}))
@@ -111,7 +116,7 @@ def art_in_path(path):
         return os.path.join(path, images[0])
 
 
-# Main interface.
+# Try each source in turn.
 
 def art_for_album(album, path):
     """Given an album info dictionary from MusicBrainz, returns a path
@@ -131,3 +136,36 @@ def art_for_album(album, path):
     else:
         log.debug('No ASIN available: no art found.')
         return None
+
+
+# PLUGIN LOGIC ###############################################################
+
+class FetchArtPlugin(BeetsPlugin):
+    def __init__(self):
+        super(FetchArtPlugin, self).__init__()
+        self.import_stages = [self.fetch_art]
+        self.register_listener('import_task_files', self.assign_art)
+
+        # Holds paths to downloaded images between fetching them and
+        # placing them in the filesystem.
+        self.art_paths = {}
+
+    # Asynchronous; after music is added to the library.
+    def fetch_art(self, config, task):
+        """Find art for the album being imported."""
+        if task.should_write_tags() and task.is_album:
+            path = art_for_album(task.info, task.path)
+            if path:
+                self.art_paths[task] = path
+
+    # Synchronous; after music files are put in place.
+    def assign_art(self, config, task):
+        """Place the discovered art in the filesystem."""
+        if task in self.art_paths:
+            path = self.art_paths.pop(task)
+
+            album = config.lib.get_album(task.album_id)
+            album.set_art(path, not (config.delete or config.move))
+
+            if config.delete or config.move:
+                task.prune(path)