+++ test/_common.py
@@ -84,7 +84,6 @@ def iconfig(lib, **kwargs):
         copy = True,
         move = False,
         write = False,
-        art = False,
         delete = False,
         choose_match_func = lambda x, y: importer.action.SKIP,
         should_resume_func = lambda _: False,